% Auto-generated: do not edit by hand
\name{dashRjsf}

\alias{dashRjsf}

\title{DashRjsf component}

\description{
ExampleComponent is an example component. It takes a property, `schema`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dashRjsf(children=NULL, id=NULL, schema=NULL, uischema=NULL,
value=NULL)
}

\arguments{
\item{children}{Logical | numeric | character | named list | unnamed list. }

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{schema}{Logical | numeric | character | named list | unnamed list. A schema that will be printed when this component is rendered.}

\item{uischema}{Logical | numeric | character | named list | unnamed list. }

\item{value}{Logical | numeric | character | named list | unnamed list. The value displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

